Clean up secrets-redaction
-Move valid_segment_traits to function group constructors
-Move secret_traits to main class constructor
-Move unit testing for additional secret redaction back into logger tests for user_admin
-add input cleanup to generate request only path and remove cleanup on entry to functions
-docstring updates and minor code improvements
