+++ pyracf/common/logger.py
@@ -100,12 +100,10 @@ class Logger:
         dictionary: dict,
         secret_traits: dict,
     ) -> dict:
-        """Redact a list of specified secret traits in a dictionary"""
+        """Redact a list of specified secret traits in a request dictionary"""
         for seg_trait in secret_traits:
             segment = seg_trait.split(":")[0]
-            if not isinstance(dictionary.get(segment, {}).get(seg_trait, {}), dict):
-                return dictionary
-            if not dictionary.get(segment, {}).get(seg_trait, {}).get("value", ""):
+            if not dictionary.get(segment, {}).get(seg_trait, {}).get("value", None):
                 continue
             dictionary[segment][seg_trait]["value"] = "********"
         return dictionary
@@ -117,6 +115,7 @@ class Logger:
         end_pattern: str,
     ):
         """Redacts characters in a string between a starting index and ending pattern"""
+        """Replaces the identified characters with ******** (regardless of length)"""
         pre_keyword = input_string[:start_ind]
         post_keyword = end_pattern.join(input_string[start_ind:].split(end_pattern)[1:])
         return pre_keyword + "********" + end_pattern + post_keyword
@@ -127,6 +126,7 @@ class Logger:
         secret_traits: dict,
     ) -> Union[str, bytes]:
         """Redact a list of specific secret traits in a request xml string or bytes object"""
+        """Bases this on xml pattern <xml_key any-operation>xml_value</xml_key>"""
         is_bytes = False
         if isinstance(xml_string, bytes):
             is_bytes = True
@@ -135,7 +135,7 @@ class Logger:
             match = re.search(rf"\<{xml_key}+[^>]*\>", xml_string)
             if not match:
                 continue
-            xml_string = self.__redact_string(xml_string, match.end(), "</")
+            xml_string = self.__redact_string(xml_string, match.end(), f"</{xml_key}")
         if is_bytes:
             xml_string = xml_string.encode("utf-8")
         return xml_string
@@ -146,6 +146,7 @@ class Logger:
         secret_traits: dict,
     ) -> str:
         """Redacts a list of specific secret traits in a result xml string"""
+        """Bases this on RACF command pattern RACF_KEY( value ) """
         for xml_key in secret_traits.values():
             racf_key = xml_key.split(":")[1] if ":" in xml_key else xml_key
             match = re.search(rf"{racf_key.upper()} +\(", xml_string)
