+++ pyracf/connection/connection_admin.py
@@ -9,27 +9,6 @@ from pyracf.connection.connection_request import ConnectionRequest
 class ConnectionAdmin(SecurityAdmin):
     """RACF Connection Administration."""
 
-    _valid_segment_traits = {
-        "base": {
-            "base:adsp": "racf:adsp",
-            "base:auditor": "racf:auditor",
-            "base:auth": "racf:auth",
-            "base:cgauthda": "racf:cgauthda",
-            "base:cginitct": "racf:cginitct",
-            "base:cgljdate": "racf:cgljdate",
-            "base:cgljtime": "racf:cgljtime",
-            "base:group": "racf:group",
-            "base:group_access": "racf:grpacc",
-            "base:operations": "racf:oper",
-            "base:owner": "racf:owner",
-            "base:resume": "racf:resume",
-            "base:revoke": "racf:revoke",
-            "base:revokefl": "racf:revokefl",
-            "base:special": "racf:special",
-            "base:uacc": "racf:uacc",
-        }
-    }
-
     def __init__(
         self,
         debug: bool = False,
@@ -38,6 +17,26 @@ class ConnectionAdmin(SecurityAdmin):
         overwrite_field_data: Union[dict, None] = None,
         additional_secret_traits: Union[List[str], None] = None,
     ) -> None:
+        self._valid_segment_traits = {
+            "base": {
+                "base:adsp": "racf:adsp",
+                "base:auditor": "racf:auditor",
+                "base:auth": "racf:auth",
+                "base:cgauthda": "racf:cgauthda",
+                "base:cginitct": "racf:cginitct",
+                "base:cgljdate": "racf:cgljdate",
+                "base:cgljtime": "racf:cgljtime",
+                "base:group": "racf:group",
+                "base:group_access": "racf:grpacc",
+                "base:operations": "racf:oper",
+                "base:owner": "racf:owner",
+                "base:resume": "racf:resume",
+                "base:revoke": "racf:revoke",
+                "base:revokefl": "racf:revokefl",
+                "base:special": "racf:special",
+                "base:uacc": "racf:uacc",
+            }
+        }
         super().__init__(
             "groupConnection",
             debug=debug,
