+++ pyracf/data_set/data_set_admin.py
@@ -10,50 +10,6 @@ from .data_set_request import DataSetRequest
 class DataSetAdmin(SecurityAdmin):
     """RACF Data Set Profile Administration."""
 
-    _valid_segment_traits = {
-        "base": {
-            "base:altvol": "racf:altvol",
-            "base:category": "racf:category",
-            "base:creatdat": "racf:creatdat",
-            "base:data": "racf:data",
-            "base:dsns": "racf:dsns",
-            "base:dstype": "racf:dstype",
-            "base:erase": "racf:erase",
-            "base:fclass": "racf:fclass",
-            "base:fgeneric": "racf:fgeneric",
-            "base:fileseq": "racf:fileseq",
-            "base:from": "racf:from",
-            "base:groupnm": "racf:groupnm",
-            "base:history": "racf:history",
-            "base:id": "racf:id",
-            "base:lchgdat": "racf:lchgdat",
-            "base:level": "racf:level",
-            "base:lrefdat": "racf:lrefdat",
-            "base:model": "racf:model",
-            "base:noracf": "racf:noracf",
-            "base:notify": "racf:notify",
-            "base:owner": "racf:owner",
-            "base:prefix": "racf:prefix",
-            "base:profile": "racf:profile",
-            "base:raudit": "racf:raudit",
-            "base:retpd": "racf:retpd",
-            "base:rgaudit": "racf:rgaudit",
-            "base:seclabel": "racf:seclabel",
-            "base:seclevel": "racf:seclevel",
-            "base:set": "racf:set",
-            "base:setonly": "racf:setonly",
-            "base:stats": "racf:stats",
-            "base:tape": "racf:tape",
-            "base:universal_access": "racf:uacc",
-            "base:unit": "racf:unit",
-            "base:volume": "racf:volume",
-            "base:volser": "racf:volser",
-            "base:warning": "racf:warning",
-        },
-        "dfp": {"dfp:resowner": "racf:resowner", "dfp:datakey": "racf:datakey"},
-        "tme": {"tme:roles": "racf:roles"},
-    }
-
     def __init__(
         self,
         debug: bool = False,
@@ -62,6 +18,49 @@ class DataSetAdmin(SecurityAdmin):
         overwrite_field_data: Union[dict, None] = None,
         additional_secret_traits: Union[List[str], None] = None,
     ) -> None:
+        self._valid_segment_traits = {
+            "base": {
+                "base:altvol": "racf:altvol",
+                "base:category": "racf:category",
+                "base:creatdat": "racf:creatdat",
+                "base:data": "racf:data",
+                "base:dsns": "racf:dsns",
+                "base:dstype": "racf:dstype",
+                "base:erase": "racf:erase",
+                "base:fclass": "racf:fclass",
+                "base:fgeneric": "racf:fgeneric",
+                "base:fileseq": "racf:fileseq",
+                "base:from": "racf:from",
+                "base:groupnm": "racf:groupnm",
+                "base:history": "racf:history",
+                "base:id": "racf:id",
+                "base:lchgdat": "racf:lchgdat",
+                "base:level": "racf:level",
+                "base:lrefdat": "racf:lrefdat",
+                "base:model": "racf:model",
+                "base:noracf": "racf:noracf",
+                "base:notify": "racf:notify",
+                "base:owner": "racf:owner",
+                "base:prefix": "racf:prefix",
+                "base:profile": "racf:profile",
+                "base:raudit": "racf:raudit",
+                "base:retpd": "racf:retpd",
+                "base:rgaudit": "racf:rgaudit",
+                "base:seclabel": "racf:seclabel",
+                "base:seclevel": "racf:seclevel",
+                "base:set": "racf:set",
+                "base:setonly": "racf:setonly",
+                "base:stats": "racf:stats",
+                "base:tape": "racf:tape",
+                "base:universal_access": "racf:uacc",
+                "base:unit": "racf:unit",
+                "base:volume": "racf:volume",
+                "base:volser": "racf:volser",
+                "base:warning": "racf:warning",
+            },
+            "dfp": {"dfp:resowner": "racf:resowner", "dfp:datakey": "racf:datakey"},
+            "tme": {"tme:roles": "racf:roles"},
+        }
         super().__init__(
             "dataSet",
             debug=debug,
