+++ pyracf/resource/resource_admin.py
@@ -10,165 +10,6 @@ from .resource_request import ResourceRequest
 class ResourceAdmin(SecurityAdmin):
     """General Resaurce Profile Administration."""
 
-    _valid_segment_traits = {
-        "base": {
-            "base:appldata": "racf:appldata",
-            "base:automatc": "racf:automatc",
-            "base:category": "racf:category",
-            "base:creatdat": "racf:creatdat",
-            "base:data": "racf:data",
-            "base:fclass": "racf:fclass",
-            "base:fgeneric": "racf:fgeneric",
-            "base:fprofile": "racf:fprofile",
-            "base:history": "racf:history",
-            "base:lchgdat": "racf:lchgdat",
-            "base:level": "racf:level",
-            "base:lrefdat": "racf:lrefdat",
-            "base:member": "racf:member",
-            "base:noracf": "racf:noracf",
-            "base:notify": "racf:notify",
-            "base:noyourac": "racf:noyourac",
-            "base:owner": "racf:owner",
-            "base:profile": "racf:profile",
-            "base:raudit": "racf:raudit",
-            "base:resgroup": "racf:resgroup",
-            "base:rgaudit": "racf:rgaudit",
-            "base:seclabel": "racf:seclabel",
-            "base:seclevel": "racf:seclevel",
-            "base:singldsn": "racf:singldsn",
-            "base:stats": "racf:stats",
-            "base:timezone": "racf:timezone",
-            "base:tvtoc": "racf:tvtoc",
-            "base:universal_access": "racf:uacc",
-            "base:volume": "racf:volume",
-            "base:warning": "racf:warning",
-            "base:whendays": "racf:whendays",
-            "base:whentime": "racf:whentime",
-        },
-        "cdtinfo": {
-            "cdtinfo:cdtcase": "case",
-            "cdtinfo:cdtdftrc": "defaultrc",
-            "cdtinfo:cdtfirst": "first",
-            "cdtinfo:cdtgen": "generic",
-            "cdtinfo:cdtgenl": "genlist",
-            "cdtinfo:cdtgroup": "grouping",
-            "cdtinfo:cdtkeyql": "keyqual",
-            "cdtinfo:cdtmac": "macprocessing",
-            "cdtinfo:cdtmaxln": "maxlenx",
-            "cdtinfo:cdtmaxlx": "maxlength",
-            "cdtinfo:cdtmembr": "member",
-            "cdtinfo:cdtoper": "operations",
-            "cdtinfo:cdtother": "other",
-            "cdtinfo:cdtposit": "posit",
-            "cdtinfo:cdtprfal": "profilesallowed",
-            "cdtinfo:cdtracl": "raclist",
-            "cdtinfo:cdtsigl": "signal",
-            "cdtinfo:cdtslreq": "seclabelrequired",
-            "cdtinfo:cdtuacc": "defaultuacc",
-        },
-        "cfdef": {
-            "cfdef:cfdtype": "type",
-            "cfdef:cffirst": "first",
-            "cfdef:cfhelp": "help",
-            "cfdef:cflist": "listhead",
-            "cfdef:cfmixed": "mixed",
-            "cfdef:cfmnval": "minvalue",
-            "cfdef:cfmxlen": "maxlength",
-            "cfdef:cfmxval": "other",
-            "cfdef:cfother": "other",
-            "cfdef:cfvalrx": "racf:cfvalrx",
-        },
-        "dlfdata": {
-            "dlfdata:jobname": "racf:jobname",
-            "dlfdata:retain": "racf:retain",
-        },
-        "eim": {
-            "eim:domaindn": "domaindn",
-            "eim:kerbreg": "kerberg",
-            "eim:localreg": "localreg",
-            "eim:options": "options",
-            "eim:x509reg": "X509reg",
-        },
-        "kerb": {
-            "kerb:chkaddrs": "checkaddrs",
-            "kerb:deftktlf": "deftktlife",
-            "kerb:encrypt": "encrypt",
-            "kerb:kerbname": "kerbname",
-            "kerb:keyvers": "racf:keyvers",
-            "kerb:maxtktlf": "maxtktlf",
-            "kerb:mintktlf": "mintklife",
-            "kerb:password": "password",
-        },
-        "icsf": {
-            "icsf:crtlbls": "symexportcert",
-            "icsf:export": "symexportable",
-            "icsf:keylbls": "symexportkey",
-            "icsf:scpwrap": "symcpacfwrap",
-            "icsf:scpret": "symcpacfret",
-            "icsf:usage": "asymusage",
-        },
-        "ictx": {
-            "ictx:domap": "domap",
-            "ictx:mapreq": "mapreq",
-            "ictx:maptimeo": "maptimeo",
-            "ictx:usemap": "usemap",
-        },
-        "idtparms": {
-            "idtpamrs:sigtoken": "sigtoken",
-            "idtparms:sigseqn": "sigseqnum",
-            "idtparms:sigcat": "sigcat",
-            "idtparms:sigalg": "sigalg",
-            "idtparms:idttimeo": "idttimeout",
-            "idtpamrs:anyappl": "anyappl",
-        },
-        "jes": {"jes:keylabel": "racf:keylabel"},
-        "mfpolicy": {
-            "mfpolicy:factors": "racf:factors",
-            "mfpolicy:timeout": "racf:timeout",
-            "mfpolicy:reuse": "racf:reuse",
-        },
-        "proxy": {
-            "proxy:binddn": "binddn",
-            "proxy:bindpw": "bindpw",
-            "proxy:ldaphost": "ldaphost",
-        },
-        "session": {
-            "session:convsec": "racf:convsec",
-            "session:interval": "racf:interval",
-            "session:lock": "racf:lock",
-            "session:sesskey": "racf:sesskey",
-        },
-        "sigver": {
-            "sigver:failload": "failload",
-            "sigver:sigaudit": "sigaudit",
-            "sigver:sigreqd": "sigrequired",
-        },
-        "ssignon": {
-            "ssigon:keycrypt": "racf:keycrypt",
-            "ssigon:ptkeylab": "ptkeylab",
-            "ssigon:pttype": "pttype",
-            "ssigon:pttimeo": "pttimeo",
-            "ssigon:ptreplay": "ptreplay",
-            "ssigon:keylabel": "racf:keylabel",
-            "ssigno:keymask": "racf:keymask",
-        },
-        "stdata": {
-            "ssigon:group": "racf:group",
-            "ssigon:privlege": "racf:privlege",
-            "ssigon:trace": "racf:trace",
-            "ssigon:trusted": "racf:trusted",
-            "ssigon:user": "racf:user",
-        },
-        "svfmr": {"svfmr:parmname": "racf:parmname", "svfmr:script": "racf:script"},
-        "tme": {
-            "tme:children": "racf:children",
-            "tme:groups": "racf:groups",
-            "tme:parent": "racf:parent",
-            "tme:resource": "racf:resource",
-            "tme:roles": "racf:roles",
-        },
-    }
-
     def __init__(
         self,
         debug: bool = False,
@@ -177,6 +18,164 @@ class ResourceAdmin(SecurityAdmin):
         overwrite_field_data: Union[dict, None] = None,
         additional_secret_traits: Union[List[str], None] = None,
     ) -> None:
+        self._valid_segment_traits = {
+            "base": {
+                "base:appldata": "racf:appldata",
+                "base:automatc": "racf:automatc",
+                "base:category": "racf:category",
+                "base:creatdat": "racf:creatdat",
+                "base:data": "racf:data",
+                "base:fclass": "racf:fclass",
+                "base:fgeneric": "racf:fgeneric",
+                "base:fprofile": "racf:fprofile",
+                "base:history": "racf:history",
+                "base:lchgdat": "racf:lchgdat",
+                "base:level": "racf:level",
+                "base:lrefdat": "racf:lrefdat",
+                "base:member": "racf:member",
+                "base:noracf": "racf:noracf",
+                "base:notify": "racf:notify",
+                "base:noyourac": "racf:noyourac",
+                "base:owner": "racf:owner",
+                "base:profile": "racf:profile",
+                "base:raudit": "racf:raudit",
+                "base:resgroup": "racf:resgroup",
+                "base:rgaudit": "racf:rgaudit",
+                "base:seclabel": "racf:seclabel",
+                "base:seclevel": "racf:seclevel",
+                "base:singldsn": "racf:singldsn",
+                "base:stats": "racf:stats",
+                "base:timezone": "racf:timezone",
+                "base:tvtoc": "racf:tvtoc",
+                "base:universal_access": "racf:uacc",
+                "base:volume": "racf:volume",
+                "base:warning": "racf:warning",
+                "base:whendays": "racf:whendays",
+                "base:whentime": "racf:whentime",
+            },
+            "cdtinfo": {
+                "cdtinfo:cdtcase": "case",
+                "cdtinfo:cdtdftrc": "defaultrc",
+                "cdtinfo:cdtfirst": "first",
+                "cdtinfo:cdtgen": "generic",
+                "cdtinfo:cdtgenl": "genlist",
+                "cdtinfo:cdtgroup": "grouping",
+                "cdtinfo:cdtkeyql": "keyqual",
+                "cdtinfo:cdtmac": "macprocessing",
+                "cdtinfo:cdtmaxln": "maxlenx",
+                "cdtinfo:cdtmaxlx": "maxlength",
+                "cdtinfo:cdtmembr": "member",
+                "cdtinfo:cdtoper": "operations",
+                "cdtinfo:cdtother": "other",
+                "cdtinfo:cdtposit": "posit",
+                "cdtinfo:cdtprfal": "profilesallowed",
+                "cdtinfo:cdtracl": "raclist",
+                "cdtinfo:cdtsigl": "signal",
+                "cdtinfo:cdtslreq": "seclabelrequired",
+                "cdtinfo:cdtuacc": "defaultuacc",
+            },
+            "cfdef": {
+                "cfdef:cfdtype": "type",
+                "cfdef:cffirst": "first",
+                "cfdef:cfhelp": "help",
+                "cfdef:cflist": "listhead",
+                "cfdef:cfmixed": "mixed",
+                "cfdef:cfmnval": "minvalue",
+                "cfdef:cfmxlen": "maxlength",
+                "cfdef:cfmxval": "other",
+                "cfdef:cfother": "other",
+                "cfdef:cfvalrx": "racf:cfvalrx",
+            },
+            "dlfdata": {
+                "dlfdata:jobname": "racf:jobname",
+                "dlfdata:retain": "racf:retain",
+            },
+            "eim": {
+                "eim:domaindn": "domaindn",
+                "eim:kerbreg": "kerberg",
+                "eim:localreg": "localreg",
+                "eim:options": "options",
+                "eim:x509reg": "X509reg",
+            },
+            "kerb": {
+                "kerb:chkaddrs": "checkaddrs",
+                "kerb:deftktlf": "deftktlife",
+                "kerb:encrypt": "encrypt",
+                "kerb:kerbname": "kerbname",
+                "kerb:keyvers": "racf:keyvers",
+                "kerb:maxtktlf": "maxtktlf",
+                "kerb:mintktlf": "mintklife",
+                "kerb:password": "password",
+            },
+            "icsf": {
+                "icsf:crtlbls": "symexportcert",
+                "icsf:export": "symexportable",
+                "icsf:keylbls": "symexportkey",
+                "icsf:scpwrap": "symcpacfwrap",
+                "icsf:scpret": "symcpacfret",
+                "icsf:usage": "asymusage",
+            },
+            "ictx": {
+                "ictx:domap": "domap",
+                "ictx:mapreq": "mapreq",
+                "ictx:maptimeo": "maptimeo",
+                "ictx:usemap": "usemap",
+            },
+            "idtparms": {
+                "idtpamrs:sigtoken": "sigtoken",
+                "idtparms:sigseqn": "sigseqnum",
+                "idtparms:sigcat": "sigcat",
+                "idtparms:sigalg": "sigalg",
+                "idtparms:idttimeo": "idttimeout",
+                "idtpamrs:anyappl": "anyappl",
+            },
+            "jes": {"jes:keylabel": "racf:keylabel"},
+            "mfpolicy": {
+                "mfpolicy:factors": "racf:factors",
+                "mfpolicy:timeout": "racf:timeout",
+                "mfpolicy:reuse": "racf:reuse",
+            },
+            "proxy": {
+                "proxy:binddn": "binddn",
+                "proxy:bindpw": "bindpw",
+                "proxy:ldaphost": "ldaphost",
+            },
+            "session": {
+                "session:convsec": "racf:convsec",
+                "session:interval": "racf:interval",
+                "session:lock": "racf:lock",
+                "session:sesskey": "racf:sesskey",
+            },
+            "sigver": {
+                "sigver:failload": "failload",
+                "sigver:sigaudit": "sigaudit",
+                "sigver:sigreqd": "sigrequired",
+            },
+            "ssignon": {
+                "ssigon:keycrypt": "racf:keycrypt",
+                "ssigon:ptkeylab": "ptkeylab",
+                "ssigon:pttype": "pttype",
+                "ssigon:pttimeo": "pttimeo",
+                "ssigon:ptreplay": "ptreplay",
+                "ssigon:keylabel": "racf:keylabel",
+                "ssigno:keymask": "racf:keymask",
+            },
+            "stdata": {
+                "ssigon:group": "racf:group",
+                "ssigon:privlege": "racf:privlege",
+                "ssigon:trace": "racf:trace",
+                "ssigon:trusted": "racf:trusted",
+                "ssigon:user": "racf:user",
+            },
+            "svfmr": {"svfmr:parmname": "racf:parmname", "svfmr:script": "racf:script"},
+            "tme": {
+                "tme:children": "racf:children",
+                "tme:groups": "racf:groups",
+                "tme:parent": "racf:parent",
+                "tme:resource": "racf:resource",
+                "tme:roles": "racf:roles",
+            },
+        }
         super().__init__(
             "resource",
             debug=debug,
