+++ pyracf/setropts/setropts_admin.py
@@ -10,93 +10,6 @@ from .setropts_requset import SetroptsRequest
 class SetroptsAdmin(SecurityAdmin):
     """Set RACF Options Administration."""
 
-    _valid_segment_traits = {
-        "base": {
-            "base:active_class": "racf:classact",
-            "base:addcreat": "racf:addcreat",
-            "base:adsp": "racf:adsp",
-            "base:applaudt": "racf:applaudt",
-            "base:audit_class": "racf:audit",
-            "base:catdsns": "racf:catdsns",
-            "base:cmdviol": "racf:cmdviol",
-            "base:compmode": "racf:compmode",
-            "base:egn": "racf:egn",
-            "base:erase": "racf:erase",
-            "base:eraseall": "racf:eraseall",
-            "base:erasesec": "racf:erasesec",
-            "base:general_command_class": "racf:gencmd",
-            "base:generic_profile_checking_class": "racf:generic",
-            "base:generic_profile_sharing_class": "racf:genlist",
-            "base:genowner": "racf:genowner",
-            "base:global_access_class": "racf:global",
-            "base:grplist": "racf:grplist",
-            "base:history": "racf:history",
-            "base:inactive": "racf:inactive",
-            "base:initstat": "racf:initstat",
-            "base:interval": "racf:interval",
-            "base:jesbatch": "racf:jesbatch",
-            "base:jesearly": "racf:jesearly",
-            "base:jesnje": "racf:jesnje",
-            "base:jesundef": "racf:jesundef",
-            "base:jesxbm": "racf:jesxbm",
-            "base:kerblvl": "racf:kerblvl",
-            "base:list": "racf:list",
-            "base:logalwys": "racf:logalwys",
-            "base:logdeflt": "racf:logdeflt",
-            "base:logfail": "racf:logfail",
-            "base:lognever": "racf:lognever",
-            "base:logsucc": "racf:logsucc",
-            "base:minchang": "racf:minchang",
-            "base:mixdcase": "racf:mixdcase",
-            "base:mlactive": "racf:mlactive",
-            "base:mlfs": "racf:mlfs",
-            "base:mlipc": "racf:mlipc",
-            "base:mlnames": "racf:mlnames",
-            "base:mlquiet": "racf:mlquiet",
-            "base:mls": "racf:mls",
-            "base:mlstable": "racf:mlstable",
-            "base:model": "racf:model",
-            "base:modgdg": "racf:modgdg",
-            "base:modgroup": "racf:modgroup",
-            "base:moduser": "racf:moduser",
-            "base:operaudt": "racf:operaudt",
-            "base:phrint": "racf:phrint",
-            "base:prefix": "racf:prefix",
-            "base:primlang": "racf:primlang",
-            "base:protall": "racf:protall",
-            "base:pwdalg": "racf:pwdalg",
-            "base:pwdspec": "racf:pwdspec",
-            "base:raclist": "racf:raclist",
-            "base:realdsn": "racf:realdsn",
-            "base:refresh": "racf:refresh",
-            "base:retpd": "racf:retpd",
-            "base:revoke": "racf:revoke",
-            "base:rules": "racf:rules",
-            "base:rule1": "racf:rule1",
-            "base:rule2": "racf:rule2",
-            "base:rule3": "racf:rule3",
-            "base:rule4": "racf:rule4",
-            "base:rule5": "racf:rule5",
-            "base:rule6": "racf:rule6",
-            "base:rule7": "racf:rule7",
-            "base:rule8": "racf:rule8",
-            "base:rvarswpw": "racf:rvarswpw",
-            "base:rvarstpw": "racf:rvarstpw",
-            "base:saudit": "racf:saudit",
-            "base:seclabct": "racf:seclabct",
-            "base:seclang": "racf:seclang",
-            "base:sessint": "racf:sessint",
-            "base:slabaudt": "racf:slabaudt",
-            "base:slbysys": "racf:slbysys",
-            "base:slevaudt": "racf:slevaudt",
-            "base:statistics_class": "racf:classtat",
-            "base:tapedsn": "racf:tapedsn",
-            "base:terminal": "racf:terminal",
-            "base:warning": "racf:warning",
-            "base:whenprog": "racf:whenprog",
-        }
-    }
-
     def __init__(
         self,
         debug: bool = False,
@@ -105,6 +18,92 @@ class SetroptsAdmin(SecurityAdmin):
         overwrite_field_data: Union[dict, None] = None,
         additional_secret_traits: Union[List[str], None] = None,
     ) -> None:
+        self._valid_segment_traits = {
+            "base": {
+                "base:active_class": "racf:classact",
+                "base:addcreat": "racf:addcreat",
+                "base:adsp": "racf:adsp",
+                "base:applaudt": "racf:applaudt",
+                "base:audit_class": "racf:audit",
+                "base:catdsns": "racf:catdsns",
+                "base:cmdviol": "racf:cmdviol",
+                "base:compmode": "racf:compmode",
+                "base:egn": "racf:egn",
+                "base:erase": "racf:erase",
+                "base:eraseall": "racf:eraseall",
+                "base:erasesec": "racf:erasesec",
+                "base:general_command_class": "racf:gencmd",
+                "base:generic_profile_checking_class": "racf:generic",
+                "base:generic_profile_sharing_class": "racf:genlist",
+                "base:genowner": "racf:genowner",
+                "base:global_access_class": "racf:global",
+                "base:grplist": "racf:grplist",
+                "base:history": "racf:history",
+                "base:inactive": "racf:inactive",
+                "base:initstat": "racf:initstat",
+                "base:interval": "racf:interval",
+                "base:jesbatch": "racf:jesbatch",
+                "base:jesearly": "racf:jesearly",
+                "base:jesnje": "racf:jesnje",
+                "base:jesundef": "racf:jesundef",
+                "base:jesxbm": "racf:jesxbm",
+                "base:kerblvl": "racf:kerblvl",
+                "base:list": "racf:list",
+                "base:logalwys": "racf:logalwys",
+                "base:logdeflt": "racf:logdeflt",
+                "base:logfail": "racf:logfail",
+                "base:lognever": "racf:lognever",
+                "base:logsucc": "racf:logsucc",
+                "base:minchang": "racf:minchang",
+                "base:mixdcase": "racf:mixdcase",
+                "base:mlactive": "racf:mlactive",
+                "base:mlfs": "racf:mlfs",
+                "base:mlipc": "racf:mlipc",
+                "base:mlnames": "racf:mlnames",
+                "base:mlquiet": "racf:mlquiet",
+                "base:mls": "racf:mls",
+                "base:mlstable": "racf:mlstable",
+                "base:model": "racf:model",
+                "base:modgdg": "racf:modgdg",
+                "base:modgroup": "racf:modgroup",
+                "base:moduser": "racf:moduser",
+                "base:operaudt": "racf:operaudt",
+                "base:phrint": "racf:phrint",
+                "base:prefix": "racf:prefix",
+                "base:primlang": "racf:primlang",
+                "base:protall": "racf:protall",
+                "base:pwdalg": "racf:pwdalg",
+                "base:pwdspec": "racf:pwdspec",
+                "base:raclist": "racf:raclist",
+                "base:realdsn": "racf:realdsn",
+                "base:refresh": "racf:refresh",
+                "base:retpd": "racf:retpd",
+                "base:revoke": "racf:revoke",
+                "base:rules": "racf:rules",
+                "base:rule1": "racf:rule1",
+                "base:rule2": "racf:rule2",
+                "base:rule3": "racf:rule3",
+                "base:rule4": "racf:rule4",
+                "base:rule5": "racf:rule5",
+                "base:rule6": "racf:rule6",
+                "base:rule7": "racf:rule7",
+                "base:rule8": "racf:rule8",
+                "base:rvarswpw": "racf:rvarswpw",
+                "base:rvarstpw": "racf:rvarstpw",
+                "base:saudit": "racf:saudit",
+                "base:seclabct": "racf:seclabct",
+                "base:seclang": "racf:seclang",
+                "base:sessint": "racf:sessint",
+                "base:slabaudt": "racf:slabaudt",
+                "base:slbysys": "racf:slbysys",
+                "base:slevaudt": "racf:slevaudt",
+                "base:statistics_class": "racf:classtat",
+                "base:tapedsn": "racf:tapedsn",
+                "base:terminal": "racf:terminal",
+                "base:warning": "racf:warning",
+                "base:whenprog": "racf:whenprog",
+            }
+        }
         super().__init__(
             "systemSettings",
             debug=debug,
