+++ pyracf/user/user_admin.py
@@ -10,199 +10,6 @@ from .user_request import UserRequest
 class UserAdmin(SecurityAdmin):
     """User Administration."""
 
-    _valid_segment_traits = {
-        "base": {
-            "base:adsp": "racf:adsp",
-            "base:auditor": "racf:auditor",
-            "base:auth": "racf:auth",
-            "base:category": "racf:category",
-            "base:class_authorizations": "racf:clauth",
-            "base:connects": "racf:connects",
-            "base:cadsp": "racf:cadsp",
-            "base:cauditor": "racf:cauditor",
-            "base:cauthda": "racf:cauthda",
-            "base:cgroup": "racf:cgroup",
-            "base:cgrpacc": "racf:cgrpacc",
-            "base:cinitct": "racf:cinitct",
-            "base:cljdate": "racf:cljdate",
-            "base:cljtime": "racf:cljtime",
-            "base:coper": "racf:coper",
-            "base:cowner": "racf:cowner",
-            "base:cresume": "racf:cresume",
-            "base:crevoke": "racf:crevoke",
-            "base:crevokfl": "racf:crevokfl",
-            "base:cspecial": "racf:cspecial",
-            "base:cuacc": "racf:cuacc",
-            "base:creatdat": "racf:creatdat",
-            "base:data": "racf:data",
-            "base:dfltgrp": "defgroup",
-            "base:expired": "racf:expired",
-            "base:factorn": "racf:factorn",
-            "base:factor": "racf:factor",
-            "base:facactv": "racf:facactv",
-            "base:factagnn": "racf:factagnn",
-            "base:facvalnn": "racf:facvalnn",
-            "base:group": "racf:group",
-            "base:grpacc": "racf:grpacc",
-            "base:hasphras": "racf:hasphras",
-            "base:haspwd": "racf:haspwd",
-            "base:lastdate": "racf:lastdate",
-            "base:lasttime": "racf:lasttime",
-            "base:mfaflbk": "racf:mfaflbk",
-            "base:mfapolnm": "racf:mfapolnm",
-            "base:model": "racf:model",
-            "base:name": "name",
-            "base:oidcard": "racf:oidcard",
-            "base:operations": "racf:oper",
-            "base:owner": "racf:owner",
-            "base:passdate": "racf:passdate",
-            "base:passint": "racf:passint",
-            "base:password": "racf:password",
-            "base:passphrase": "racf:phrase",
-            "base:phrdate": "racf:phrdate",
-            "base:phrint": "racf:phrint",
-            "base:pphenv": "racf:pphenv",
-            "base:protectd": "racf:protectd",
-            "base:pwdenv": "racf:pwdenv",
-            "base:rest": "racf:rest",
-            "base:resume": "resumedate",
-            "base:revoke": "revokedate",
-            "base:revokefl": "racf:revokefl",
-            "base:roaudit": "racf:roaudit",
-            "base:seclabel": "seclabel",
-            "base:seclevel": "racf:seclevel",
-            "base:special": "racf:special",
-            "base:uacc": "racf:uacc",
-            "base:uaudit": "uaudit",
-            "base:whendays": "whendays",
-            "base:whensrv": "whensrv",
-            "base:whentime": "whentime",
-        },
-        "cics": {
-            "cisc:opclass": "racf:opclass",
-            "cics:opident": "opident",
-            "cics:opprty": "opprty",
-            "cics:rslkey": "racf:rslkey",
-            "cics:timeout": "timeout",
-            "cics:tslkey": "racf:tslkey",
-            "cics:xrfsoff": "force",
-        },
-        "dce": {
-            "dce:autolog": "autolog",
-            "dce:dcename": "dcename",
-            "dce:homecell": "homecell",
-            "dce:homeuuid": "homeuuid",
-            "dce:uuid": "uuid",
-        },
-        "dfp": {
-            "dfp:dataappl": "dataappl",
-            "dfp:dataclas": "dataclas",
-            "dfp:mgmtclas": "mgmtclass",
-            "dfp:storclas": "storclas",
-        },
-        "eim": {"ldapprof": "racf:ldapprof"},
-        "kerb": {
-            "dfp:encrypt": "racf:encrypt",
-            "dfp:kerbname": "racf:kerbname",
-            "dfp:keyfrom": "racf:keyfrom",
-            "dfp:keyvers": "racf:keyvers",
-            "dfp:maxtktlf": "racf:maxtktlf",
-        },
-        "language": {"language:primary": "primary", "language:second": "secondary"},
-        "lnotes": {"lnotes:sname": "racf:sname"},
-        "mfa": {
-            "mfa:factor": "racf:factor",
-            "mfa:facactv": "racf:facactv",
-            "mfa:factags": "racf:factags",
-            "mfa:mfaflbk": "racf:mfaflbk",
-            "mfa:mfapolnm": "racf:mfapolnm",
-        },
-        "nds": {"nds:uname": "racf:uname"},
-        "netview": {
-            "netview:consname": "consid",
-            "netview:ctl": "secctl",
-            "netview:domains": "nvdomains",
-            "netview:ic": "ic",
-            "netview:msgrecvr": "msgrec",
-            "netview:ngmfadmn": "racf:ngmfadmn",
-            "netview:ngmfvspn": "gmfadmin",
-            "netview:opclass": "racf:opclass",
-        },
-        "omvs": {
-            "omvs:assize": "assize",
-            "omvs:autouid": "racf:autouid",
-            "omvs:cputime": "cputime",
-            "omvs:fileproc": "filemax",
-            "omvs:home": "home",
-            "omvs:memlimit": "memlim",
-            "omvs:mmaparea": "mmaparea",
-            "omvs:procuser": "procmax",
-            "omvs:program": "pgm",
-            "omvs:shared": "racf:shared",
-            "omvs:shmemmax": "shmemmax",
-            "omvs:threads": "threads",
-            "omvs:uid": "uid",
-        },
-        "operparm": {
-            "operparm:altgrp": "altgrp",
-            "operparm:auto": "auto",
-            "operparm:cmdsys": "cmdsys",
-            "operparm:dom": "dom",
-            "operparm:hc": "hc",
-            "operparm:intids": "intid",
-            "operparm:key": "key",
-            "operparm:level": "racf:level",
-            "operparm:logcmd": "logcmd",
-            "operparm:mform": "mform",
-            "operparm:migid": "migid",
-            "operparm:monitor": "mon",
-            "operparm:mscope": "racf:mscope",
-            "operparm:operauth": "auth",
-            "operparm:routcode": "routcode",
-            "operparm:storage": "storage",
-            "operparm:ud": "ud",
-            "operparm:unknids": "unkids",
-        },
-        "ovm": {
-            "ovm:fsroot": "racf:fsroot",
-            "ovm:vhome": "racf:vhome",
-            "ovm:vprogram": "racf:vprogram",
-            "ovm:vuid": "racf:vuid",
-        },
-        "proxy": {
-            "proxy:binddn": "racf:binddn",
-            "proxy:bindpw": "racf:bindpw",
-            "proxy:ldaphost": "racf:ldaphost",
-        },
-        "tso": {
-            "tso:acctnum": "acctnum",
-            "tso:command": "command",
-            "tso:dest": "dest",
-            "tso:hldclass": "holdclass",
-            "tso:jobclass": "jobclass",
-            "tso:maxsize": "maxsize",
-            "tso:msgclass": "msgclass",
-            "tso:proc": "proc",
-            "tso:seclabel": "seclabel",
-            "tso:size": "size",
-            "tso:sysoutcl": "sysclass",
-            "tso:unit": "unit",
-            "tso:userdata": "userdata",
-        },
-        "workattr": {
-            "workattr:waaccnt": "waaccnt",
-            "workattr:waaddr1": "waaddr1",
-            "workattr:waaddr2": "waaddr2",
-            "workattr:waaddr3": "waaddr3",
-            "workattr:waaddr4": "waaddr4",
-            "workattr:wabldg": "wabldg",
-            "workattr:wadept": "wadept",
-            "workattr:waname": "waname",
-            "workattr:waroom": "waroom",
-            "workattr:waemail": "waemail",
-        },
-    }
-
     def __init__(
         self,
         debug: bool = False,
@@ -211,6 +18,198 @@ class UserAdmin(SecurityAdmin):
         overwrite_field_data: Union[dict, None] = None,
         additional_secret_traits: Union[List[str], None] = None,
     ) -> None:
+        self._valid_segment_traits = {
+            "base": {
+                "base:adsp": "racf:adsp",
+                "base:auditor": "racf:auditor",
+                "base:auth": "racf:auth",
+                "base:category": "racf:category",
+                "base:class_authorizations": "racf:clauth",
+                "base:connects": "racf:connects",
+                "base:cadsp": "racf:cadsp",
+                "base:cauditor": "racf:cauditor",
+                "base:cauthda": "racf:cauthda",
+                "base:cgroup": "racf:cgroup",
+                "base:cgrpacc": "racf:cgrpacc",
+                "base:cinitct": "racf:cinitct",
+                "base:cljdate": "racf:cljdate",
+                "base:cljtime": "racf:cljtime",
+                "base:coper": "racf:coper",
+                "base:cowner": "racf:cowner",
+                "base:cresume": "racf:cresume",
+                "base:crevoke": "racf:crevoke",
+                "base:crevokfl": "racf:crevokfl",
+                "base:cspecial": "racf:cspecial",
+                "base:cuacc": "racf:cuacc",
+                "base:creatdat": "racf:creatdat",
+                "base:data": "racf:data",
+                "base:dfltgrp": "defgroup",
+                "base:expired": "racf:expired",
+                "base:factorn": "racf:factorn",
+                "base:factor": "racf:factor",
+                "base:facactv": "racf:facactv",
+                "base:factagnn": "racf:factagnn",
+                "base:facvalnn": "racf:facvalnn",
+                "base:group": "racf:group",
+                "base:grpacc": "racf:grpacc",
+                "base:hasphras": "racf:hasphras",
+                "base:haspwd": "racf:haspwd",
+                "base:lastdate": "racf:lastdate",
+                "base:lasttime": "racf:lasttime",
+                "base:mfaflbk": "racf:mfaflbk",
+                "base:mfapolnm": "racf:mfapolnm",
+                "base:model": "racf:model",
+                "base:name": "name",
+                "base:oidcard": "racf:oidcard",
+                "base:operations": "racf:oper",
+                "base:owner": "racf:owner",
+                "base:passdate": "racf:passdate",
+                "base:passint": "racf:passint",
+                "base:password": "racf:password",
+                "base:passphrase": "racf:phrase",
+                "base:phrdate": "racf:phrdate",
+                "base:phrint": "racf:phrint",
+                "base:pphenv": "racf:pphenv",
+                "base:protectd": "racf:protectd",
+                "base:pwdenv": "racf:pwdenv",
+                "base:rest": "racf:rest",
+                "base:resume": "resumedate",
+                "base:revoke": "revokedate",
+                "base:revokefl": "racf:revokefl",
+                "base:roaudit": "racf:roaudit",
+                "base:seclabel": "seclabel",
+                "base:seclevel": "racf:seclevel",
+                "base:special": "racf:special",
+                "base:uacc": "racf:uacc",
+                "base:uaudit": "uaudit",
+                "base:whendays": "whendays",
+                "base:whensrv": "whensrv",
+                "base:whentime": "whentime",
+            },
+            "cics": {
+                "cisc:opclass": "racf:opclass",
+                "cics:opident": "opident",
+                "cics:opprty": "opprty",
+                "cics:rslkey": "racf:rslkey",
+                "cics:timeout": "timeout",
+                "cics:tslkey": "racf:tslkey",
+                "cics:xrfsoff": "force",
+            },
+            "dce": {
+                "dce:autolog": "autolog",
+                "dce:dcename": "dcename",
+                "dce:homecell": "homecell",
+                "dce:homeuuid": "homeuuid",
+                "dce:uuid": "uuid",
+            },
+            "dfp": {
+                "dfp:dataappl": "dataappl",
+                "dfp:dataclas": "dataclas",
+                "dfp:mgmtclas": "mgmtclass",
+                "dfp:storclas": "storclas",
+            },
+            "eim": {"ldapprof": "racf:ldapprof"},
+            "kerb": {
+                "dfp:encrypt": "racf:encrypt",
+                "dfp:kerbname": "racf:kerbname",
+                "dfp:keyfrom": "racf:keyfrom",
+                "dfp:keyvers": "racf:keyvers",
+                "dfp:maxtktlf": "racf:maxtktlf",
+            },
+            "language": {"language:primary": "primary", "language:second": "secondary"},
+            "lnotes": {"lnotes:sname": "racf:sname"},
+            "mfa": {
+                "mfa:factor": "racf:factor",
+                "mfa:facactv": "racf:facactv",
+                "mfa:factags": "racf:factags",
+                "mfa:mfaflbk": "racf:mfaflbk",
+                "mfa:mfapolnm": "racf:mfapolnm",
+            },
+            "nds": {"nds:uname": "racf:uname"},
+            "netview": {
+                "netview:consname": "consid",
+                "netview:ctl": "secctl",
+                "netview:domains": "nvdomains",
+                "netview:ic": "ic",
+                "netview:msgrecvr": "msgrec",
+                "netview:ngmfadmn": "racf:ngmfadmn",
+                "netview:ngmfvspn": "gmfadmin",
+                "netview:opclass": "racf:opclass",
+            },
+            "omvs": {
+                "omvs:assize": "assize",
+                "omvs:autouid": "racf:autouid",
+                "omvs:cputime": "cputime",
+                "omvs:fileproc": "filemax",
+                "omvs:home": "home",
+                "omvs:memlimit": "memlim",
+                "omvs:mmaparea": "mmaparea",
+                "omvs:procuser": "procmax",
+                "omvs:program": "pgm",
+                "omvs:shared": "racf:shared",
+                "omvs:shmemmax": "shmemmax",
+                "omvs:threads": "threads",
+                "omvs:uid": "uid",
+            },
+            "operparm": {
+                "operparm:altgrp": "altgrp",
+                "operparm:auto": "auto",
+                "operparm:cmdsys": "cmdsys",
+                "operparm:dom": "dom",
+                "operparm:hc": "hc",
+                "operparm:intids": "intid",
+                "operparm:key": "key",
+                "operparm:level": "racf:level",
+                "operparm:logcmd": "logcmd",
+                "operparm:mform": "mform",
+                "operparm:migid": "migid",
+                "operparm:monitor": "mon",
+                "operparm:mscope": "racf:mscope",
+                "operparm:operauth": "auth",
+                "operparm:routcode": "routcode",
+                "operparm:storage": "storage",
+                "operparm:ud": "ud",
+                "operparm:unknids": "unkids",
+            },
+            "ovm": {
+                "ovm:fsroot": "racf:fsroot",
+                "ovm:vhome": "racf:vhome",
+                "ovm:vprogram": "racf:vprogram",
+                "ovm:vuid": "racf:vuid",
+            },
+            "proxy": {
+                "proxy:binddn": "racf:binddn",
+                "proxy:bindpw": "racf:bindpw",
+                "proxy:ldaphost": "racf:ldaphost",
+            },
+            "tso": {
+                "tso:acctnum": "acctnum",
+                "tso:command": "command",
+                "tso:dest": "dest",
+                "tso:hldclass": "holdclass",
+                "tso:jobclass": "jobclass",
+                "tso:maxsize": "maxsize",
+                "tso:msgclass": "msgclass",
+                "tso:proc": "proc",
+                "tso:seclabel": "seclabel",
+                "tso:size": "size",
+                "tso:sysoutcl": "sysclass",
+                "tso:unit": "unit",
+                "tso:userdata": "userdata",
+            },
+            "workattr": {
+                "workattr:waaccnt": "waaccnt",
+                "workattr:waaddr1": "waaddr1",
+                "workattr:waaddr2": "waaddr2",
+                "workattr:waaddr3": "waaddr3",
+                "workattr:waaddr4": "waaddr4",
+                "workattr:wabldg": "wabldg",
+                "workattr:wadept": "wadept",
+                "workattr:waname": "waname",
+                "workattr:waroom": "waroom",
+                "workattr:waemail": "waemail",
+            },
+        }
         super().__init__(
             "user",
             debug=debug,
