+++ tests/test_runner.py
@@ -36,9 +36,6 @@ from tests.setropts.test_setropts_result_parser import TestSetroptsResultParser
 from tests.setropts.test_setropts_setters import TestSetroptsSetters
 from tests.user.test_user_debug_logging import TestUserDebugLogging
 from tests.user.test_user_getters import TestUserGetters
-from tests.user.test_user_redact_additional_secrets import (
-    TestUserRedactAdditionalSecrets,
-)
 from tests.user.test_user_request_builder import TestUserRequestBuilder
 from tests.user.test_user_result_parser import TestUserResultParser
 from tests.user.test_user_setters import TestUserSetters
@@ -84,7 +81,6 @@ def __test_suite() -> unittest.TestSuite:
         TestUserGetters,
         TestUserSetters,
         TestUserDebugLogging,
-        TestUserRedactAdditionalSecrets,  # Will break if this is not the last user test case run
     ]
     for test_class in test_classes:
         tests = test_loader.loadTestsFromTestCase(test_class)
