+++ tests/user/test_user_debug_logging.py
@@ -184,6 +184,54 @@ class TestUserDebugLogging(unittest.TestCase):
         self.assertNotIn(self.test_passphrase, error_log)
         self.assertNotIn(self.test_password, error_log)
 
+    def test_add_user_request_debug_log_added_secrets_get_redacted_on_success(
+        self,
+        call_racf_mock: Mock,
+    ):
+        self.user_admin_local = UserAdmin(
+            debug=True, additional_secret_traits=["omvs:uid"]
+        )
+        call_racf_mock.return_value = TestUserConstants.TEST_ADD_USER_RESULT_SUCCESS_XML
+        stdout = io.StringIO()
+        with contextlib.redirect_stdout(stdout):
+            self.user_admin_local.add(
+                "squidwrd",
+                traits=TestUserConstants.TEST_ADD_USER_REQUEST_TRAITS,
+            )
+        success_log = self.ansi_escape.sub("", stdout.getvalue())
+        self.assertEqual(
+            success_log,
+            TestUserConstants.TEST_ADD_USER_ADDED_SECRET_SUCCESS_LOG,
+        )
+        self.assertNotIn(
+            TestUserConstants.TEST_ADD_USER_REQUEST_TRAITS["omvs:uid"], success_log
+        )
+
+    def test_add_user_request_debug_log_added_secrets_get_redacted_on_error(
+        self,
+        call_racf_mock: Mock,
+    ):
+        self.user_admin_local = UserAdmin(
+            debug=True, additional_secret_traits=["omvs:uid"]
+        )
+        call_racf_mock.return_value = TestUserConstants.TEST_ADD_USER_RESULT_ERROR_XML
+        stdout = io.StringIO()
+        with contextlib.redirect_stdout(stdout):
+            try:
+                self.user_admin_local.add(
+                    "squidwrd",
+                    traits=TestUserConstants.TEST_ADD_USER_REQUEST_TRAITS,
+                )
+            except SecurityRequestError:
+                pass
+        error_log = self.ansi_escape.sub("", stdout.getvalue())
+        self.assertEqual(
+            error_log, TestUserConstants.TEST_ADD_USER_ADDED_SECRET_ERROR_LOG
+        )
+        self.assertNotIn(
+            TestUserConstants.TEST_ADD_USER_REQUEST_TRAITS["omvs:uid"], error_log
+        )
+
     # ============================================================================
     # Extract User
     # ============================================================================
